Sjednocene typy a pridan displej
Vsechny typy sjednoceny na uin8_t apod
Pridan output pro nokia displej

+++ AlwaysCondition.cpp
@@ -10,7 +10,7 @@
 
 
 
-bool AlwaysCondition::check( int value )
+bool AlwaysCondition::check( int16_t value )
 {
 	
 }
